refactor(web): extract right drawer panel ids
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
