+++ apps/web/src/shared/types/drawer.ts
@@ -1,0 +1,8 @@
+export type DrawerPanelId =
+  | 'properties'
+  | 'validation'
+  | 'connections'
+  | 'scenarios'
+  | 'learning'
+  | 'code'
+  | 'templates';
